Move make_optimizer_and_scheduler to utils
This allows users to use this functionality from their scripts.
